+++ dev.js
@@ -1,6 +1,8 @@
 require('dotenv').config();
 const app = require('./index');
 
-app.listen(3000, () => {
-	console.log('server started on port 3000');
+const port = process.env.PORT || 8000;
+
+app.listen(port, () => {
+	console.log(`Server started on port: ${port}...`);
 })
