+++ index.js
@@ -3,7 +3,7 @@ const cors = require('cors');
 const { PokemonService } = require('./service');
 const swaggerUI = require('swagger-ui-express');
 const swaggerDoc = require('./swagger.json');
-const { PokemonController } = require('./controller');
+const { PokemonController } = require('./controller/controller');
 
 const pokemonController = new PokemonController(new PokemonService());
